Added all bonus functions

+++ libft/ft_strlen.c
@@ -6,17 +6,15 @@
 /*   By: pamallet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 14:19:36 by pamallet          #+#    #+#             */
-/*   Updated: 2024/10/16 13:05:28 by pamallet         ###   ########.fr       */
+/*   Updated: 2024/10/31 11:29:59 by pamallet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
-#include <string.h>
 #include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	size_t		i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i])
